requirements.txt
Django==1.8.3
django-filter==0.10.0
djangorestframework==3.1.3
wheel==0.24.0
